pyproject.toml
[tool.poetry]
name = "w800btplus-for-linux"
version = "0.1.0"
description = "Um API para controle e monitoramento do Headset EDIFIER W800BT Plus para Linux mais especificamente no Ubuntu."
authors = ["Victor Cunha <cunhavictorr@gmail.com>"]
readme = "README.md"
packages = [{include = "w800btplus_for_linux"}]

[tool.poetry.scripts]
app-w800btplus = "w800btplus_for_linux.cli:app"

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.7.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
pytest-cov = "^4.0.0"
black = "^23.1.0"
isort = "^5.12.0"
taskipy = "^1.10.3"

[tool.poetry.group.doc.dependencies]
mkdocs = "^1.4.2"
mkdocstrings = "^0.20.0"
mkdocstrings-python = "^0.8.3"
mkdocs-material = "^9.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "--doctest-modules"

[tool.isort]
profile = "black"
line_length = 79

[tool.black]
line_length = 79

[tool.taskipy.tasks]
lint = "black --diff . && isort --diff ."
docs = "mkdocs serve"
pre_test = "task lint"
test = "pytest -s -x --cov=w800btplus_for_linux -vv"
post_test = "coverage html"